separated feature collection definition from impl helpers
