+++ datatypes/src/collections/mod.rs
@@ -1,8 +1,9 @@
-#[macro_use]
-mod feature_collection;
 pub(self) mod error;
+mod feature_collection;
 #[macro_use]
 mod geo_feature_collection;
+#[macro_use]
+mod feature_collection_impl;
 
 mod data_collection;
 mod multi_line_collection;
@@ -11,7 +12,7 @@ mod multi_polygon_collection;
 
 pub(crate) use error::FeatureCollectionError;
 pub use feature_collection::FeatureCollection;
-pub(self) use feature_collection::FeatureCollectionImplHelpers;
+pub(self) use feature_collection_impl::FeatureCollectionImplHelpers;
 pub use geo_feature_collection::{IntoGeometryIterator, IntoGeometryOptionsIterator};
 
 pub use data_collection::DataCollection;
